Improve usage example code in readme
The latest version of xBRjs expects a canvas with an image already drawn on it, which is much better compared to previous versions that expected an <img>. However, the current usage example code in the readme can be confusing, as it assumes you have `canvas` (with an image drawn on it), `context`, `sourceWidth` and `sourceHeight` variables.

I split the example code up in two sections, the first section is optional and explains how to turn an <img> into a canvas, and the second section shows how to apply xbr2x to an existing canvas. By doing so, this example code should help both those who have an <img>, and those who have an existing canvas.

Furthermore, I eliminated `scaledWidth` and `scaledHeight` and managed to keep the example code concise like it was before.

Lastly, I replaced `new ImageData` with `context.createImageData`, because the former is not available in all environments. See the browser compatibility table for the ImageData constructor for more information: https://developer.mozilla.org/en-US/docs/Web/API/ImageData